SkillSample refactor to TestBotClient
